+++ device_storage.py
@@ -1,3 +1,6 @@
+import Domoticz
+
+
 class DeviceStorage:
     # Here will be the instance stored.
     __instance = None
@@ -24,6 +27,8 @@ class DeviceStorage:
                 'ieee_addr': item['ieeeAddr'],
                 'friendly_name': item['friendly_name']
             }
+
+            Domoticz.Debug('Device ' + item['model'] + ' ' + item['ieeeAddr'] + ' (' + item['friendly_name'] + ')')
 
     def get_device_by_id(self, ieee_addr):
         return self.devices[ieee_addr] if ieee_addr in self.devices else None